linux-firmware: update to 20240513
Update the linux-firmware recipe to the most recent upstream tag.

LICENSE.amdgpu md5 checksum changed because they bumped the Copyright
year[1].

Some of qcom-sdm845-modem files got replaced by a symlink to their
counterpart in ath10k/WCN3990/hw1.0/. The same happened to files in
qcom-qcm2290-wifi, but to ath10k/WCN3990/hw1.0/qcm2290/. Finally,
qcom-qrb4210-wifi also got its file symlinked to
ath10k/WCN3990/hw1.0/qrb4210/ which in turn is a symlink to
ath10k/WCN3990/hw1.0/qcm2290/.[2] This meant that qrb4210-wifi now
RDEPENDS on qcm2290-wifi for the target of the links to be present when
installing it, and also that sdm845-modem and qcm2290-wifi packages have
to be before the ath10k catch-all package.

[1] https://git.kernel.org/pub/scm/linux/kernel/git/firmware/linux-firmware.git/commit/?id=9442c72e361c1e1130a19248598ad6ab5282ba09
[2] https://git.kernel.org/pub/scm/linux/kernel/git/firmware/linux-firmware.git/commit/?id=650e88378e76d5fad3997a5398f1ade47a74d924
Signed-off-by: Quentin Schulz <quentin.schulz@cherry.de>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>