libtool: 2.4.7 -> 2.5.0
Whilst this is an alpha release, it makes sense for us to
stay close to upstream.

We can drop a lot of merged patches which is great to see.

During testing, one bug showed up, particularly on mip64 due
to an issue with FILECMD being changed to use AC_CHECK_PROG
incorredly. A patch has been added for that and sent upstream.

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>