pciutils: rewrite recipe
The upstream Makefiles are bespoke so need to be handheld, but this
recipe has workarounds for code removed a decade ago.

Add more PACKAGECONFIGs for the other configuration options (zlib, kmod),
and use hwdb by default (non-systemd builds also have hwdb, as it's part
of eudev too).

Instead of using 'make -e', pass the variables that we want the Makefile
to respect explicitly (CC, CFLAGS, etc).

Construct an appropriate HOST so that lib/configure does the right thing
without needing patches.

This now results in text relocations (apparently from the versioned
symbols in libpci), so add an INSANE_SKIP for that.

Update the homepage URL.

Signed-off-by: Ross Burton <ross.burton@arm.com>
Signed-off-by: Alexandre Belloni <alexandre.belloni@bootlin.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>