findutils: upgrade from 4.9.0 to 4.10.0
autoconf-2.73.patch is dropped as it's not needed any more.

The run-ptest script is fixed. Instead copying function definitions
from init.cfg file, we should just install init.cfg file in the ptest
package. The coreutils-getlimits is added as a runtime dependency of
the ptest package because the getlimits command is needed.

Signed-off-by: Chen Qi <Qi.Chen@windriver.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>