abi_version/package: Bump hashequiv version and package class version
The recent pkgconfig change and subsqeuent fixes have left the hash equivalence
server in a corrupted state with hashes linking the changes beofore and after
the pkgconfig change, breaking reproducibile builds.

Bump the appropriate version numbers to allow us to move on and avoid
build failures and corrupt equivalence data now the underlying issue
was fixed.

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>