meson: Fix native meson config
When the optimization flags were reduced to a single option, a different
and clearly untested codepath was used where the output was unquoted. This
caused errors when testing meson in the SDK.

Use repr() in the single item case as per the multiple item case to correctly
quote the output and fix build failures.

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>