icu: Backport patch to fix build issues with long paths (>512 chars)
There is a bug libicu that causes libicu to be installed incorrectly when
the build system uses long paths (more than 512 chars).

This condition is not very difficult to trigger on a OE build system
due to the long paths an the deep of the directories that are usually
generated by default.
Also the bug is very subtle and won't be detected by the QA post-install
processes because what this bug causes is that a different version of
libicudata.so.X.Y (one without data) is installed instead of the one
containing the data, but there won't be any file missed on the installation
(just that it installed the wrong one).

See: https://unicode-org.atlassian.net/browse/ICU-22813

This patch backports the fix from upstream/main

Signed-off-by: Carlos Alberto Lopez Perez <clopez@igalia.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>