insane: Add missing vardepsexclude
If we improve the function dependency visibility in insane.bbclass, it
exposes some dependencies which were previously not seen causing variances
in the do_package_qa task checksums. Update vardepsexclude in a couple of
test cases to ensure the sstate hash selftests pass and the taskhashes
don't vary when we don't expect them too.

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>